# specs/openapi.yaml
openapi: 3.0.3
info:
  title: eyeO Platform API
  description: |
    Privacy-preserving video surveillance platform with Zero-Trust architecture.
    
    ## Architecture Overview
    The eyeO platform implements a Shared-Nothing Architecture (SNA) with split-brain data flows:
    - **Red Flow**: Protected video data (client-side processing only)
    - **Blue Flow**: Detection events and metadata (server-side analytics)
    
    ## Security Model
    - JWT-based authentication (HS512)
    - Client-side data transformation
    - Role-based access control (FREE/PRO/ENTERPRISE)
    - Rate limiting: 10 req/sec per user
    
    ## Base URL
    Production: `https://api.eyeo-platform.com`
    Development: `https://localhost`
    
  version: 1.0.0
  contact:
    name: eyeO Platform Security Team
    email: security@eyeo-platform.com
  license:
    name: Proprietary
    url: https://eyeo-platform.com/license
servers:
  - url: https://api.eyeo-platform.com/v1
    description: Production API
  - url: https://staging-api.eyeo-platform.com/v1
    description: Staging API
  - url: https://localhost/v1
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/api/auth/login` endpoint.
        Include in Authorization header: `Bearer <token>`
    
  # ==================== Reusable Schemas ====================
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: demo@eyeo-platform.com
          minLength: 3
          maxLength: 50
        password:
          type: string
          format: password
          example: SecurePassword123!
          minLength: 8
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzUxMiJ9...
        refreshToken:
          type: string
          description: Refresh token for obtaining new access tokens
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 86400
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: demo@eyeo-platform.com
        role:
          type: string
          enum: [FREE, PRO, ENTERPRISE]
          example: PRO
        licenseStatus:
          type: string
          enum: [ACTIVE, TRIAL, EXPIRED, SUSPENDED]
          example: ACTIVE
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: newuser
        email:
          type: string
          format: email
          example: newuser@example.com
        password:
          type: string
          format: password
          minLength: 8
        tier:
          type: string
          enum: [FREE, PRO, ENTERPRISE]
          default: FREE
    
    # Data Protection Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Invalid credentials
        timestamp:
          type: integer
          format: int64
          example: 1704235200000
        path:
          type: string
          example: /api/auth/login
    
    # Health Schema
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Validates credentials and returns JWT access token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Creates new user account with FREE tier by default
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/introspect:
    post:
      tags:
        - Authentication
      summary: Validate JWT token
      description: Used by downstream services to verify token validity
      operationId: introspectToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/UserInfo'
        '401':
          description: Invalid or expired token
  
  # ==================== Data Protection Endpoints ====================
...
  /storage/{storageKey}:
    get:
      tags:
        - Storage
      summary: Retrieve protected blob
      description: Returns protected blob by storage key. Client must process data locally.
      operationId: getProtectedBlob
      security:
        - bearerAuth: []
      parameters:
        - name: storageKey
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier for the protected blob
      responses:
        '200':
          description: Protected blob retrieved
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Blob not found
        '403':
          description: Access denied (not owner or delegate)
  
  # ==================== Event Endpoints (Blue Flow) ====================
...
security:
  - bearerAuth: []